Do not crash when loading undefined context from hash
When loading `from_hash` and with `strict_mode false`,
a config context that is not defined will cause
mixlib-config to crash.

This is because we used `internal_get` without checking
to see if the result was nil.  nil occurs when the
scenario above is met.  Now, we will define the context
when this happens, and continue to populate that context
from the hash.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>